Aggiunto slider per scegliere la lunghezza del video Aggiunta nel menu l'impostazione per scegliere se tenere i frame del video salvati Aggiunta nel menu l'impostazione per cambiare il nome del file video prodotto

+++ src/JavaFXVersion/sorting/BubbleSort.java
@@ -76,10 +76,10 @@ public class BubbleSort implements SortAlgorithm {
                             countSwaps++;
                             SortUtils.swap(array, j, j + 1);
                             swapped = true;
-                            if((countSwaps % delay) == 0)
+                            if ((countSwaps % delay) == 0)
                                 writeImage(userSettings, array, width, height, imageIndex++, countComparison, countSwaps);
 
-                            progressBar.setProgress(progress+=increment);
+                            progressBar.setProgress(progress += increment);
                         }
                     }
                     if (!swapped) {
@@ -89,9 +89,10 @@ public class BubbleSort implements SortAlgorithm {
             }
             writeImage(userSettings, array, width, height, imageIndex, countComparison, countSwaps);
             FFMPEG prc = new FFMPEG(userSettings, progressBar);
-            deleteAllPreviousFiles(userSettings);
+            if (userSettings.saveImage == false)
+                deleteAllPreviousFiles(userSettings);
 
-            if(userSettings.isOpenFile()) {
+            if (userSettings.isOpenFile()) {
                 File out = new File(userSettings.getOutputDirectory() + "\\" + userSettings.getOutName());
                 try {
                     Desktop.getDesktop().open(out);
@@ -101,7 +102,7 @@ public class BubbleSort implements SortAlgorithm {
             }
             Platform.runLater(() -> {
                 gridPane.setVisible(true);
-                ((BorderPane)gridPane.getParent()).getChildren().remove(progressBar);
+                ((BorderPane) gridPane.getParent()).getChildren().remove(progressBar);
             });
 
         });
@@ -110,7 +111,7 @@ public class BubbleSort implements SortAlgorithm {
 
     private void calculateNumberOfSwaps(Tail[] array) {
         Tail[] tmp = new Tail[array.length];
-        System.arraycopy(array,0,tmp, 0, array.length);
+        System.arraycopy(array, 0, tmp, 0, array.length);
 
         for (int size = tmp.length, i = 1; i < size; ++i) {
             boolean swapped = false;
@@ -131,16 +132,16 @@ public class BubbleSort implements SortAlgorithm {
     private void setupEnv(GridPane gridPane) {
         progressBar = new ProgressBar(0);
         increment = 1d / countSwaps;
-        ((BorderPane)gridPane.getParent()).setBottom(progressBar);
+        ((BorderPane) gridPane.getParent()).setBottom(progressBar);
         gridPane.setVisible(false);
         progressBar.setPrefWidth(gridPane.getWidth());
         progressBar.setMinWidth(gridPane.getWidth());
         progressBar.setPrefHeight(50);
         progressBar.setMinHeight(50);
         BorderPane.setAlignment(progressBar, Pos.CENTER);
-        BorderPane.setMargin(progressBar, new Insets(0,0,10,0));
+        BorderPane.setMargin(progressBar, new Insets(0, 0, 10, 0));
 
-        delay = countSwaps / (userSettings.getFrameRate() * 15) + 1;
+        delay = countSwaps / (userSettings.getFrameRate() * userSettings.getVideoDuration()) + 1;
     }
 
     @Override
