+++ src/JavaFXVersion/sorting/GnomeSort.java
@@ -58,7 +58,7 @@ public class GnomeSort implements SortAlgorithm {
             int i = 1;
             int n = array.length;
             while (i < n) {
-                if(running == false)
+                if (running == false)
                     break;
                 ++countComparison;
                 if (i == 0 || SortUtils.greater(array[i], array[i - 1])) {
@@ -70,14 +70,15 @@ public class GnomeSort implements SortAlgorithm {
                     array[--i] = tmp;
                     if ((countSwaps % delay) == 0)
                         writeImage(userSettings, array, width, height, imageIndex++, countComparison, countSwaps);
-                    progressBar.setProgress(progress+=increment);
+                    progressBar.setProgress(progress += increment);
                 }
             }
             writeImage(userSettings, array, width, height, imageIndex, countComparison, countSwaps);
             FFMPEG prc = new FFMPEG(userSettings, progressBar);
-            deleteAllPreviousFiles(userSettings);
+            if (userSettings.saveImage == false)
+                deleteAllPreviousFiles(userSettings);
 
-            if(userSettings.isOpenFile()) {
+            if (userSettings.isOpenFile()) {
                 File out = new File(userSettings.getOutputDirectory() + "\\" + userSettings.getOutName());
                 try {
                     Desktop.getDesktop().open(out);
@@ -87,7 +88,7 @@ public class GnomeSort implements SortAlgorithm {
             }
             Platform.runLater(() -> {
                 gridPane.setVisible(true);
-                ((BorderPane)gridPane.getParent()).getChildren().remove(progressBar);
+                ((BorderPane) gridPane.getParent()).getChildren().remove(progressBar);
             });
         });
         thread.start();
@@ -96,16 +97,16 @@ public class GnomeSort implements SortAlgorithm {
     private void setupEnv(GridPane gridPane) {
         progressBar = new ProgressBar(0);
         increment = 1d / countSwaps;
-        ((BorderPane)gridPane.getParent()).setBottom(progressBar);
+        ((BorderPane) gridPane.getParent()).setBottom(progressBar);
         gridPane.setVisible(false);
         progressBar.setPrefWidth(gridPane.getWidth());
         progressBar.setMinWidth(gridPane.getWidth());
         progressBar.setPrefHeight(50);
         progressBar.setMinHeight(50);
         BorderPane.setAlignment(progressBar, Pos.CENTER);
-        BorderPane.setMargin(progressBar, new Insets(0,0,10,0));
+        BorderPane.setMargin(progressBar, new Insets(0, 0, 10, 0));
 
-        delay = countSwaps / (userSettings.getFrameRate() * 15) + 1;
+        delay = countSwaps / (userSettings.getFrameRate() * userSettings.getVideoDuration()) + 1;
     }
 
     @Override
@@ -117,7 +118,7 @@ public class GnomeSort implements SortAlgorithm {
     private void calculateNumberOfSwaps(Tail[] array) {
 
         Tail[] tmp = new Tail[array.length];
-        System.arraycopy(array,0,tmp, 0, array.length);
+        System.arraycopy(array, 0, tmp, 0, array.length);
         int i = 1;
         int n = tmp.length;
         while (i < n) {
