+++ src/JavaFXVersion/sorting/QuickSort.java
@@ -37,6 +37,7 @@ public class QuickSort implements SortAlgorithm {
     private ProgressBar progressBar;
     private double progress = 0;
     private double increment;
+
     public QuickSort(UserSettings userSettings) {
         this.userSettings = userSettings;
     }
@@ -63,11 +64,12 @@ public class QuickSort implements SortAlgorithm {
         height = (int) (array[0].getImage().getHeight() % 2 == 0 ? array[0].getImage().getHeight() :
                 array[0].getImage().getHeight() - 1);
         thread = new Thread(() -> {
-            doSort(array, 0, array.length - 1, gridPane,true);
+            doSort(array, 0, array.length - 1, gridPane, true);
             FFMPEG prc = new FFMPEG(userSettings, progressBar);
-            deleteAllPreviousFiles(userSettings);
+            if (userSettings.saveImage == false)
+                deleteAllPreviousFiles(userSettings);
 
-            if(userSettings.isOpenFile()) {
+            if (userSettings.isOpenFile()) {
                 File out = new File(userSettings.getOutputDirectory() + "\\" + userSettings.getOutName());
                 try {
                     Desktop.getDesktop().open(out);
@@ -77,7 +79,7 @@ public class QuickSort implements SortAlgorithm {
             }
             Platform.runLater(() -> {
                 gridPane.setVisible(true);
-                ((BorderPane)gridPane.getParent()).getChildren().remove(progressBar);
+                ((BorderPane) gridPane.getParent()).getChildren().remove(progressBar);
             });
         });
         thread.start();
@@ -86,23 +88,23 @@ public class QuickSort implements SortAlgorithm {
     private void setupEnv(GridPane gridPane) {
         progressBar = new ProgressBar(0);
         increment = 1d / countSwaps;
-        ((BorderPane)gridPane.getParent()).setBottom(progressBar);
+        ((BorderPane) gridPane.getParent()).setBottom(progressBar);
         gridPane.setVisible(false);
         progressBar.setPrefWidth(gridPane.getWidth());
         progressBar.setMinWidth(gridPane.getWidth());
         progressBar.setPrefHeight(50);
         progressBar.setMinHeight(50);
         BorderPane.setAlignment(progressBar, Pos.CENTER);
-        BorderPane.setMargin(progressBar, new Insets(0,0,10,0));
+        BorderPane.setMargin(progressBar, new Insets(0, 0, 10, 0));
 
-        delay = countSwaps / (userSettings.getFrameRate() * 15) + 1;
+        delay = countSwaps / (userSettings.getFrameRate() * userSettings.getVideoDuration()) + 1;
     }
 
 
     private void calculateNumberOfSwaps(Tail[] array, GridPane gridPane) {
         Tail[] tmp = new Tail[array.length];
         System.arraycopy(array, 0, tmp, 0, array.length);
-        doSort(tmp, 0, tmp.length - 1, gridPane,false);
+        doSort(tmp, 0, tmp.length - 1, gridPane, false);
     }
 
     /**
@@ -140,9 +142,9 @@ public class QuickSort implements SortAlgorithm {
         int randomIndex = left + (int) (Math.random() * (right - left + 1));
 
         countSwaps++;
-        if((countSwaps % delay) == 0 && write)
+        if ((countSwaps % delay) == 0 && write)
             writeImage(userSettings, array, width, height, imageIndex++, countComparison, countSwaps);
-        if(write)
+        if (write)
             progressBar.setProgress(progress += increment);
         swap(array, randomIndex, right);
